Split implementations to different files. [#CLICKHOUSE-2070]
